Updated Address DAO

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/entity/RestaurantEntity.java
@@ -2,7 +2,6 @@ package com.upgrad.FoodOrderingApp.service.entity;
 
 import javax.persistence.*;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -99,7 +98,7 @@ public class RestaurantEntity {
         this.photo_url = photo_url;
     }
 
-    public BigDecimal getCustomer_rating() {
+    public Double getCustomer_rating() {
         return customer_rating;
     }
